chore(deployment): sync ~/.hermes/ runtime artifacts into fork
Brings the deployed files under version control so further edits can be
tracked via git instead of hand-editing ~/.hermes/ blind.

Changes:
- dashboard/monitor.py → dashboard/hermes-monitor-mac.py (rename to match deployed)
  * 179 → 350 lines (tunnel health check + auto-resync from deployed copy)
  * Today's fix: fail-streak counter → cloudflared self-restart on origin stale
  * Today's fix: gateway PID change detection → auto tunnel kickstart (60s cooldown)

- dashboard/api.py: sync with ~/.hermes (fc-rag endpoints drifted 19 lines)
- dashboard/index.html: sync with ~/.hermes (7 line drift)

- scripts/run-tunnel.sh (NEW): cloudflared supervisor
  * Today's fix: health interval 60s → 30s
  * Today's fix: FAIL_STREAK ≥ 2 → kill cloudflared, outer loop restarts session
    (worker-resync-only was ineffective against origin-stale state)

- scripts/sync-cloudflare-worker.sh (NEW): worker URL sync helper
- scripts/start-hermes-tunnel.sh (NEW): legacy bootstrap (kept for reference)

Runtime location reminder: launchd plists reference ~/.hermes/ paths,
so after pulling this branch the files must still be copied/symlinked into
~/.hermes/{dashboard,scripts}/ for the changes to take effect.
